refactored tabs and stack navigations
