+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/CustomBackButton/styles.ts
