+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/CustomHeader/index.tsx
