+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/CustomHeader/styles.ts
