+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/CustomSearchButton/index.tsx
