+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/CustomSearchButton/styles.ts
