+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/OtherScreen/styles.ts
