+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/TabbarIcon/index.tsx
