+++ src/app/components/AppNavigator/components/LibraryTabScreen/components/TabbarIcon/styles.ts
