+++ src/app/components/AppNavigator/components/LibraryTabScreen/constants/constants.ts
@@ -1,10 +1,10 @@
 import Routes from 'src/constants/routes';
-import LibraryStackScreen from '../components/LibraryStackScreen';
 import OtherScreen from '../components/OtherScreen';
+import Books from 'src/app/screens/Books';
 
 export const TABS = [Routes.Library, Routes.Wishlist, Routes.AddNew, Routes.Rentals, Routes.Settings];
 export const SCREENS = {
-  [Routes.Library]: LibraryStackScreen,
+  [Routes.Library]: Books,
   [Routes.Wishlist]: OtherScreen,
   [Routes.AddNew]: OtherScreen,
   [Routes.Rentals]: OtherScreen,
